Simplify monitor
The monitor has been simplified so that each monitor instance can be configured
with different criteria and listings for a given instance are stored in the table.
Previously the monitor would store listings by make in the table with each make being a pk.
Now the monitor name is the pk which makes having different monitors for different criteria more intuitive
and simplifies the monitor itself.

* refactoring of monitor logic
* use file path for dependencies while in testing
* update gitignore